requirements.txt
bert_score==0.3.13
chromadb==0.4.14
clip_interrogator==0.6.0
datasets==2.14.6
git+https://github.com/huggingface/diffusers
hpsv2==1.1.2
image_reward==1.5
matplotlib==3.8.0
git+https://github.com/openai/CLIP.git
pandas==2.1.2
Pillow==10.1.0
PyYAML==6.0
seaborn==0.13.0
timm==0.6.13
torch==2.1.0
tqdm==4.66.1
transformers==4.34.1
